Modularize keymappings utils
