+++ lua/user/keymappings/init.lua
@@ -1,27 +1,5 @@
-local M = {}
-
 vim.g.mapleader = " "
 vim.g.maplocalleader = " "
-
-local generic_opts_any = { noremap = true, silent = true }
-
-local generic_opts = {
-  insert_mode = generic_opts_any,
-  normal_mode = generic_opts_any,
-  visual_mode = generic_opts_any,
-  visual_block_mode = generic_opts_any,
-  command_mode = generic_opts_any,
-  term_mode = { silent = true },
-}
-
-local mode_adapters = {
-  insert_mode = "i",
-  normal_mode = "n",
-  term_mode = "t",
-  visual_mode = "v",
-  visual_block_mode = "x",
-  command_mode = "c",
-}
 
 local keymaps = {
   insert_mode = {
@@ -112,33 +90,8 @@ local keymaps = {
   },
 }
 
--- Load key mappings for a given mode
--- @param mode The keymap mode, can be one of the keys of mode_adapters
--- @param keymaps The list of key mappings
-function M.load_mode(mode, mode_keymaps)
-  mode = mode_adapters[mode] or mode
-  for k, v in pairs(mode_keymaps) do
-    M.set_keymaps(mode, k, v)
-  end
-end
-
--- Set key mappings individually
--- @param mode The keymap mode, can be one of the keys of mode_adapters
--- @param key The key of keymap
--- @param val Can be form as a mapping or tuple of mapping and user defined opt
-function M.set_keymaps(mode, key, val)
-  local opt = generic_opts[mode] or generic_opts_any
-  if type(val) == "table" then
-    opt = val[2]
-    val = val[1]
-  end
-  if val then
-    vim.keymap.set(mode, key, val, opt)
-  else
-    pcall(vim.api.nvim_del_keymap, mode, key)
-  end
-end
+local keymappings = require('user.keymappings.utils')
 
 for mode, mapping in pairs(keymaps) do
-  M.load_mode(mode, mapping)
+  keymappings.load_mode(mode, mapping)
 end
